+++ src/examples/lighting/index.ts
@@ -33,7 +33,7 @@ const Canvas3DPresets = {
                         radius: 5,
                         bias: 0.8,
                         blurKernelSize: 15,
-                        blurBias: 0.5,
+                        blurDepthBias: 0.5,
                         resolutionScale: 1,
                         color: Color(0x000000),
                     }